requirements/development.txt
wagtail>=1.9,<1.10
beautifulsoup4>=4.4.1,<4.5.0
django-extensions
ipdb
werkzeug
